Lesson 22.2: now the variables are all double (floating point) and the calculation is correct.

+++ Udemy - C Programming/main.c
@@ -2,9 +2,9 @@
 
 int main(int argc, char **argv)
 {
-	int subtotal;
-	int tax;
-	int grandtotal;
+	double subtotal;
+	double tax;
+	double grandtotal;
 	double taxrate;
 
 	taxrate = 0.175;
@@ -12,6 +12,6 @@ int main(int argc, char **argv)
 	tax = subtotal * taxrate;
 	grandtotal = subtotal + tax;
 
-	printf("The tax on %d is %d, so the grand total is %d.\n", subtotal, tax, grandtotal);
+	printf("The tax on %.2f is %.2f, so the grand total is %.2f.\n", subtotal, tax, grandtotal);
 	return 0;
 }
